:recycle: [refactory]: Update Directory
Update Directory (user of pages.)
