+++ src/pages/users/application/custom_hooks/index.ts
@@ -1,0 +1,1 @@
+export { useUserFetch } from './useUserFetch';
