+++ src/pages/users/application/domain/user.ts
@@ -1,6 +1,6 @@
 import { AxiosResponse } from 'axios';
 
-import Https from '../../../service/https';
+import Https from '../../../../service/https';
 
 interface IUser {
   selectDatas: () => Promise<AxiosResponse>;
