+++ src/pages/users/page/UserBody/index.ts
@@ -1,0 +1,1 @@
+export { default as UserBody } from './UserBody';
