+++ src/pages/users/page/index.ts
@@ -1,0 +1,1 @@
+export { default as User } from './User';
